pyproject.toml
[project]
name = "panchAga-calendar"
version = "0.1.0"
description = "Panchanga Calendar"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "icalendar>=6.1.0",
    "pydantic>=2.10.5",
    "pyquery>=2.0.1",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
    "typer>=0.15.1",
]

[dependency-groups]
dev = ["black>=24.10.0", "shubhlipi>=0.0.0.182"]
